Well names of plate type

Lists the well names of a known well plate format, from left to right and top to bottom.

Usage

PlateTypeToWellNames(type = c("6", "96", "384"))

Arguments

type

The name a plate type (character)

Value

Returns a character vector of well names.

See also

Examples

PlateTypeToWellNames("6")
#> [1] "A01" "A02" "A03" "B01" "B02" "B03"
PlateTypeToWellNames("96")
#>  [1] "A01" "A02" "A03" "A04" "A05" "A06" "A07" "A08" "A09" "A10" "A11" "A12"
#> [13] "B01" "B02" "B03" "B04" "B05" "B06" "B07" "B08" "B09" "B10" "B11" "B12"
#> [25] "C01" "C02" "C03" "C04" "C05" "C06" "C07" "C08" "C09" "C10" "C11" "C12"
#> [37] "D01" "D02" "D03" "D04" "D05" "D06" "D07" "D08" "D09" "D10" "D11" "D12"
#> [49] "E01" "E02" "E03" "E04" "E05" "E06" "E07" "E08" "E09" "E10" "E11" "E12"
#> [61] "F01" "F02" "F03" "F04" "F05" "F06" "F07" "F08" "F09" "F10" "F11" "F12"
#> [73] "G01" "G02" "G03" "G04" "G05" "G06" "G07" "G08" "G09" "G10" "G11" "G12"
#> [85] "H01" "H02" "H03" "H04" "H05" "H06" "H07" "H08" "H09" "H10" "H11" "H12"
PlateTypeToWellNames("384")
#>   [1] "A01" "A02" "A03" "A04" "A05" "A06" "A07" "A08" "A09" "A10" "A11" "A12"
#>  [13] "A13" "A14" "A15" "A16" "A17" "A18" "A19" "A20" "A21" "A22" "A23" "A24"
#>  [25] "B01" "B02" "B03" "B04" "B05" "B06" "B07" "B08" "B09" "B10" "B11" "B12"
#>  [37] "B13" "B14" "B15" "B16" "B17" "B18" "B19" "B20" "B21" "B22" "B23" "B24"
#>  [49] "C01" "C02" "C03" "C04" "C05" "C06" "C07" "C08" "C09" "C10" "C11" "C12"
#>  [61] "C13" "C14" "C15" "C16" "C17" "C18" "C19" "C20" "C21" "C22" "C23" "C24"
#>  [73] "D01" "D02" "D03" "D04" "D05" "D06" "D07" "D08" "D09" "D10" "D11" "D12"
#>  [85] "D13" "D14" "D15" "D16" "D17" "D18" "D19" "D20" "D21" "D22" "D23" "D24"
#>  [97] "E01" "E02" "E03" "E04" "E05" "E06" "E07" "E08" "E09" "E10" "E11" "E12"
#> [109] "E13" "E14" "E15" "E16" "E17" "E18" "E19" "E20" "E21" "E22" "E23" "E24"
#> [121] "F01" "F02" "F03" "F04" "F05" "F06" "F07" "F08" "F09" "F10" "F11" "F12"
#> [133] "F13" "F14" "F15" "F16" "F17" "F18" "F19" "F20" "F21" "F22" "F23" "F24"
#> [145] "G01" "G02" "G03" "G04" "G05" "G06" "G07" "G08" "G09" "G10" "G11" "G12"
#> [157] "G13" "G14" "G15" "G16" "G17" "G18" "G19" "G20" "G21" "G22" "G23" "G24"
#> [169] "H01" "H02" "H03" "H04" "H05" "H06" "H07" "H08" "H09" "H10" "H11" "H12"
#> [181] "H13" "H14" "H15" "H16" "H17" "H18" "H19" "H20" "H21" "H22" "H23" "H24"
#> [193] "I01" "I02" "I03" "I04" "I05" "I06" "I07" "I08" "I09" "I10" "I11" "I12"
#> [205] "I13" "I14" "I15" "I16" "I17" "I18" "I19" "I20" "I21" "I22" "I23" "I24"
#> [217] "J01" "J02" "J03" "J04" "J05" "J06" "J07" "J08" "J09" "J10" "J11" "J12"
#> [229] "J13" "J14" "J15" "J16" "J17" "J18" "J19" "J20" "J21" "J22" "J23" "J24"
#> [241] "K01" "K02" "K03" "K04" "K05" "K06" "K07" "K08" "K09" "K10" "K11" "K12"
#> [253] "K13" "K14" "K15" "K16" "K17" "K18" "K19" "K20" "K21" "K22" "K23" "K24"
#> [265] "L01" "L02" "L03" "L04" "L05" "L06" "L07" "L08" "L09" "L10" "L11" "L12"
#> [277] "L13" "L14" "L15" "L16" "L17" "L18" "L19" "L20" "L21" "L22" "L23" "L24"
#> [289] "M01" "M02" "M03" "M04" "M05" "M06" "M07" "M08" "M09" "M10" "M11" "M12"
#> [301] "M13" "M14" "M15" "M16" "M17" "M18" "M19" "M20" "M21" "M22" "M23" "M24"
#> [313] "N01" "N02" "N03" "N04" "N05" "N06" "N07" "N08" "N09" "N10" "N11" "N12"
#> [325] "N13" "N14" "N15" "N16" "N17" "N18" "N19" "N20" "N21" "N22" "N23" "N24"
#> [337] "O01" "O02" "O03" "O04" "O05" "O06" "O07" "O08" "O09" "O10" "O11" "O12"
#> [349] "O13" "O14" "O15" "O16" "O17" "O18" "O19" "O20" "O21" "O22" "O23" "O24"
#> [361] "P01" "P02" "P03" "P04" "P05" "P06" "P07" "P08" "P09" "P10" "P11" "P12"
#> [373] "P13" "P14" "P15" "P16" "P17" "P18" "P19" "P20" "P21" "P22" "P23" "P24"