Compute melting temperature from raw data

Although the qPCR machines output melting temperatures in their report files, it is also possible to compute them using exported raw data. This function automates running qpcR::meltcurve() for the computation in a quiet way.

Usage

calc_qPCR_TM(mlt, cut.Area = 1)

Arguments

mlt

A table providing the raw data loaded by load_qPCR_meltcurve().

cut.Area

Parameter passed directly to qpcR::meltcurve().

Value

A vector of melting temperatures in the same order as the wells of the mlt table.

See also