Titration curve

Given a number of steps and a target maximal volume, this function will return an increasing serie of volumes that cover the range in an approximately logarithmic progression. Each volume can be dispensed, as they are guaranteed to be multiple of the droplet's volume. The range can be forced to start at zero if a negative control is needed.

Usage

titration_curve(steps = 12, maxvol = 1000, dropvolume = 2.5, zero = TRUE)

Arguments

steps

The number of points in the titration curve.

maxvol

The maximum volume to reach.

dropvolume

The volume of a droplet, in nanoliters.

zero

Start with a negative control with a volume of zero.

Value

A numeric vector of length steps indicating the volumes to transfer in order to cover a volume range in a logarithmic manner.

See also

Author

Charles Plessy

Examples

titration_curve(12)
#>  [1]    0.0    2.5    5.0   10.0   17.5   32.5   55.0  100.0  177.5  315.0
#> [11]  562.5 1000.0
titration_curve(12, zero = FALSE)
#>  [1]    2.5    5.0    7.5   15.0   25.0   40.0   70.0  120.0  202.5  345.0
#> [11]  587.5 1000.0
titration_curve(12, dropvolume = 25)
#>  [1]    0   25   50   75  100  125  175  250  350  500  700 1000
titration_curve(12, dropvolume = 25, maxvol = 300)
#>  [1]   0  25  50  75 100 125 150 175 200 225 250 300