Transfer volume from a source to a destination plate

Reduce the volume of a reagent in the source plate at the coordinates of the transducer, add that volume to the destination plate, and register the transfer in the log.

Usage

transfer(echo, volume)

# S4 method for Echo,numeric
transfer(echo, volume)

Arguments

echo

An Echo object.

volume

Volume to transfer.

The what parameter must correspond to a column name in the source plate.

Value

Returns An Echo object with the source, destination and log

slots updated.

See also

Examples

sourcePlate <- examplePlate
head(sourcePlate)
#> DataFrame with 6 rows and 2 columns
#>          dNTP    buffer
#>     <numeric> <numeric>
#> A01     1e+05        NA
#> A02     1e+05        NA
#> A03        NA     1e+05
#> A04        NA        NA
#> A05        NA        NA
#> A06        NA        NA

destinationPlate <- Plate("96", deadVolume = 1e4, maxVolume = 1e5)

transducer <- Transducer( source      = Well("A01")
                        , destination = Well("A01", plateFormat = "96"))

echo <- Echo( source      = sourcePlate
            , destination = destinationPlate
            , transducer  = transducer
            , model       = "525")

echo |> showLogs()
#> data frame with 0 columns and 0 rows

echo <- echo |>
  transfer(10) |>
  transfer(10) |>
  moveTransducer(destination = Well("A02", plateFormat = "96")) |>
  transfer(5)

echo |> showLogs()
#>   from  to vol what
#> 1  A01 A01  10 dNTP
#> 2  A01 A01  10 dNTP
#> 3  A01 A02   5 dNTP